pyproject.toml
[project]
name = "text-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "gradio>=6.9.0",
    "groq>=1.1.1",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
]
